replace old tests
